refactor: split timer.rs and gc/tests/copying.rs off the 2000-line size gate
This PR pushed both files past the cap (timer.rs 1998 -> 2163, copying.rs
1939 -> 2040), failing lint. Extract the incremental GC root-scan machinery
into timer/gc_scan.rs (timer.rs is now 1947) and the side-table rewrite tests
into gc/tests/copying_side_tables.rs (copying.rs is now 1817).
